Update renderscene.js
working on renderScene drawScene. Is at the point where we are generating a list of lines with the points in each line transformed and clipped(?) not sure about clipped. From here we need to multiply each point by mper and draw. 
Sorry about gitProblems, I screwed something up and do not know what